get-by-ID functions can't abort execution anymore
This commit represents a change in the ability of the get-by-ID
functions (that is, get_name_of_id, get_genes_of_id and get_parents_of_id),
to call exit() when the ID they are given does not exist.
The change is: they no longer do that. Instead, new 'failed' return values
have been added to documentation and supplant the exit call.

This change is so that behaviour when printing pedigree and encountering
a deleted parent is more reasonable: an error is printed but the deleted
parent is treated as unknown parentage rather than causing printing to abort.