Attempt at non-macro simulation parameters
An attempt was made to create simulation parameters as a struct that
is to be a parameter in SimData.

However, removal of the CONTIGW macro means that AlleleMatrix
can no longer contain arrays but must use dynamic memory, which requires
a lot of rewriting. Therefore we'll go back to macros.