Tidy parameter ordering & remove double warning
Make it so the cap parameters come directly after the n_crosses parameters in both cross_random_individuals and cross_randomly_between. Remove warning in shuffle_up_to that was added in last commit because it's a low-level helper and it parent functions both already warn in the case tested.