Fix bug in condensing AlleleMatrix list
Fix a bug in large simulations where gaps in the AlleleMatrix linked list failed to be patched in `condense_allele_matrix`, and some genotypes lost their group numbers (group numbers reverted to 0).

Under-the-hood changes to `condense_allele_matrix`, for easier debugging. It now takes advantage of a `GappyIterator` to help it shuffle the positions of genotypes in the AlleleMatrix linked list.