Resolve some compilation warnings
The warnings were to do with passing a const SimData to a BidirectionalIterator constructor.

Couldn't figure out a way to give a BidirectionalIterator a way to promise it won't modify the simulation (without changing the simplicity of the data access functions), so the 'solution' being committed here is just trust. May revisit this problem in future.