Improve make_genotypes parameter store
Improve internal datastore used to pass parameters to the make_genotypes scaffold. (Instead of a freeform void*, it is now a union gsc_datastore_make_genotypes*)

Also, the indexes passed to make_targeted_crosses are now unsigned ints rather than ints, to better match the indexes that are used throughout the rest of the simulation. Further work on this point may be upcoming.