Fix iterator segfault when there's no genotypes
Fixed a segfault that occured when trying to create iterators for a simulation object with no loaded genotypes. This situation can occur if you load a map to set up the simulation (and so have no AlleleMatrix/no loaded genotypes at all), and then try to run some kind of data access or progression command. Those commands should simply not return anything/not do anything.